--- .github/workflows/deployment.yml
@@ -1,17 +1,0 @@
-name: Deployment
-on:
-  push:
-    branches:
-      - master
-
-jobs:
-  deploy:
-    runs-on: ubuntu-20.04
-    if: "!contains(github.event.head_commit.message, 'ci skip:')" # If the commit contains "ci skip" this action won't run
-    steps:
-    - uses: actions/checkout@v2
-    - uses: superfly/flyctl-actions@1.1
-      with:
-        args: "deploy"
-      env:
-          FLY_API_TOKEN: ${{ secrets.FLY_API_TOKEN }}
