--- Dockerfile
@@ -1,32 +1,0 @@
-# Select NodeJS LTS as image
-FROM node:lts
-
-# Create a temporary folder to build the website
-WORKDIR /tmp
-COPY package.json .
-COPY package-lock.json .
-
-# Install packages
-RUN npm ci
-
-# Copy the remaining files
-COPY . .
-
-# Build the website
-RUN npm run build --if-present
-
-# Switch to the app directory
-WORKDIR /app
-
-# Copy the build files and remove the temp folder
-RUN cp -r /tmp/build/ build/ && \
-    rm -rf /tmp
-
-# Copy the webserver
-COPY web/index.js .
-
-# Install webserver, following https://create-react-app.dev/docs/deployment
-RUN npm i express
-
-# Use port 8080 as requested: https://fly.io/docs/getting-started/node/
-CMD ["node", "index.js"]
