+++ components/NavigationBar.module.css
@@ -1,10 +1,10 @@
-.logo-image {
+.logoImage {
   overflow: hidden;
   max-height: 50px;
 }
 
 /* This is for navigation bar */
-.color-nav {
+.colorNav {
   overflow: hidden;
   background-color: rgba(255, 255, 255, 0.5);
   padding: 3px 15px;
@@ -21,11 +21,11 @@
   line-height: 20px;
 }
 
-.nav-link {
+.navLink {
   border-radius: 4px;
 }
 
-.nav-link:hover {
+.navLink:hover {
   background-color: #ddd;
   color: black;
   box-shadow: inset 0px -5px #27c7ff;
