+++ package.json
@@ -1,44 +1,19 @@
 {
-  "name": "hlresistance-preview.github.io",
+  "name": "hlresistance-frontend",
   "version": "0.1.0",
   "private": true,
-  "homepage": "https://frontend-preview.hlresistance.me/",
-  "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
-    "bootstrap": "^4.5.0",
-    "express": "^4.17.1",
-    "react": "^16.13.1",
-    "react-bootstrap": "^1.0.1",
-    "react-dom": "^16.13.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.1"
-  },
   "scripts": {
-    "start": "react-scripts start",
-    "predeploy": "npm run build",
-    "deploy": "gh-pages -d build",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "export": "next export"
   },
-  "eslintConfig": {
-    "extends": "react-app"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "devDependencies": {
-    "gh-pages": "^3.1.0"
+  "dependencies": {
+    "next": "9.4.4",
+    "next-images": "^1.4.0",
+    "react": "16.13.1",
+    "react-bootstrap": "^1.2.2",
+    "react-dom": "16.13.1",
+    "react-router-dom": "^5.2.0"
   }
 }
