+++ pages/about.js
@@ -1,0 +1,8 @@
+//filler content
+export default function About() {
+    return (
+        <div>
+            <h2>Hello about</h2>
+        </div>
+    )
+}
