+++ pages/community.js
@@ -1,0 +1,8 @@
+//filler content
+export default function Community() {
+    return (
+        <div>
+            <h2>Hello contact</h2>
+        </div>
+    )
+}
