+++ pages/creative.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import Layout from "../components/Layout";
+
+//filler content
+export default function Creative() {
+    return (
+        <Layout>
+            <div>
+                <h2>Hello creative</h2>
+            </div>
+        </Layout>
+    )
+}
