+++ pages/index.js
@@ -1,0 +1,8 @@
+//filler content
+export default function Home() {
+    return (
+        <div>
+            <h2>Hello home</h2>
+        </div>
+    )
+}
