+++ pages/translation.js
@@ -1,0 +1,8 @@
+//filler content
+export default function Translation() {
+    return (
+        <div>
+            <h2>Hello Translation</h2>
+        </div>
+    )
+}
