--- src/Community.js
@@ -1,8 +1,0 @@
-import React from 'react'
-
-//filler content
-export const Community = () => (
-    <div>
-        <h2>Hello Contact</h2>
-    </div>
-)
