--- src/Creative.js
@@ -1,8 +1,0 @@
-import React from 'react'
-
-//filler content
-export const Creative = () => (
-    <div>
-        <h2>Hello creative</h2>
-    </div>
-)
