--- src/Hololive.js
@@ -1,8 +1,0 @@
-import React from 'react'
-
-//filler content
-export const About = () => (
-    <div>
-        <h2>Hello About</h2>
-    </div>
-)
