--- src/Home.js
@@ -1,8 +1,0 @@
-import React from 'react'
-
-//filler content
-export const Home = () => (
-    <div>
-        <h2>Hello home</h2>
-    </div>
-)
