--- src/NoMatch.js
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export const NoMatch = () => (
-    <div>
-        <h2>NoMatch</h2>
-    </div>
-)
