--- src/Translation.js
@@ -1,8 +1,0 @@
-import React from 'react'
-
-//filler content
-export const Translation = () => (
-    <div>
-        <h2>Hello Translation</h2>
-    </div>
-)
